+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 UNRELEASED
 
+- Implement UserCherryPick RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/457
 - Use grpc-go 1.8.0
   https://gitlab.com/gitlab-org/gitaly/merge_requests/466
 - Fix a panic in ListFiles RPC when git process is killed abruptly