+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Gitaly changelog
 
-UNRELEASED
+v0.56.0
 
 - Implement UserCherryPick RPC
   https://gitlab.com/gitlab-org/gitaly/merge_requests/457